Merge r5865, r5866, r5867, r5870, r5871, r5872, and r5881 from Nmap Umit branch. Remove .dmp files from share/umit/config. There are files of the same names in share/umit/misc that are more up to date. The ones in config are not mentioned in the manifest.
Remove umit.db from the windows and macosx MANIFEST.in files. Also don't try to
copy umit.db when creating the user config directory. Running as a non-root
user made the copy fail, which caused an attempt to modify the umit.db in
/usr/share/umit/config, which caused a crash. Running as root worked because
the copy of umit.db (implemented by reading the contents then writing them
again) doesn't open the /usr/share database read-only, which creates a
zero-byte file which is then successfully copied. Files should probably be
opened read-only when doing these copies to avoid anomalies like this.

In copy_config_file, open the copied file in read-only mode to avoid accidentally creating it.

Simplify the code that loads icons and make it platform-independent.

Unify the three MANIFEST.in files now that they can be identical.

Install SVG icons on all platforms in the same way. I'm going to try to find the common parts of the setup.pys so that they might be combined.



git-svn-id: http://svn.umitproject.org/svnroot/umit/trunk@2617 1105ee14-b0fa-0310-85a5-ff3eed429ff7