"Comparing results against a loaded scan does nothing", this is a bug I opened earlier today and then started tracing the root of the problem. And found the cause at collect_umit_info (a method of ScanNotebookPage). This method uses the following to retrieve the attribute nmap_output from a scan:
try:
   self.parsed.nmap_output = self.command_execution.get_raw_output()
except:
   self.parsed.nmap_output = self.scan_result.get_nmap_output()

The problem here is that it will raise an exception in the try block
(because this scan was loaded from a file and the attribute
command_execution isn't created), and then will set nmap_output to the
result of self.scan_result.get_nmap_output() which is empty because
set_nmap_output is only called for scans executed during umit
execution.
The good news is that for a scan loaded from a file, nmap_output is
automatically filled in after it is parsed. So we can change those
lines to:

if hasattr(self, "command_execution"):
   self.parsed.nmap_output = self.command_execution.get_raw_output()
elif not self.parsed.nmap_output:
   self.parsed.nmap_output = self.scan_result.get_nmap_output()

Thanks to Guilherme who found and fixed this bug. The commit log is his report msg.

Also, fixed a minor english error on dates at search options. Error reported by Bartosz.

git-svn-id: http://svn.umitproject.org/svnroot/umit/trunk@2886 1105ee14-b0fa-0310-85a5-ff3eed429ff7