Fixing a bug in Mac OS X system related to Network Inventory.
Reproduce steps:

1) Tools -> Inventory -> Network Inventory
2) File -> New Invetory 
3) Press button "Use host discovery"
4) Press button "Detect network"


It will get a traceback saying "102: OSError socket doesn't support this operation".
I read about get IPs and Netmask in Mac OS and there is a function to get it getifaddrs
(man getifaddrs in BSD machines).

There is no wrapper in Python native, so what was used ctypes to wrapper it.

This function tryto_getaddresses should live in a common place because it can be useful for other project.
Unfortunaly the new structure just will apply after 1.0 release. So just for the record it should be moved to something like "umit.common".

It was added a test to have sure that IP exists and it works with loopback.



git-svn-id: http://svn.umitproject.org/svnroot/umit/trunk@5256 1105ee14-b0fa-0310-85a5-ff3eed429ff7