Removing manifest file once is not needed anymore. This approach was working on Windows Vista and Windows 7. Unfortunately it wasn't work with Windows XP+ giving an error when try execute Umit.exe: missing MSVCP90.DLL. If it was there there is another error: R6034 and it was related with Python 2.6: http://bugs.python.org/issue4566.
Create an explicit manifest file is not a good idea indeed.

It was some examples on py2exe(mainstream) showing examples of how to require access level. This commit was inspired on examples of Mark Hammond (he provides a solution to avoid this troubles and missing dll error. The manifest is embeeded and it is provided by API in py2exe.

Furthermore information, read the code committed in : http://py2exe.svn.sourceforge.net/viewvc/py2exe?view=rev&revision=670

git-svn-id: http://svn.umitproject.org/svnroot/umit/trunk@5320 1105ee14-b0fa-0310-85a5-ff3eed429ff7