Once umit is packed as tarball it doesn't have plugins-temp, nor plugins-download. If umit already was runned as sudo Umit will not have permission to run without admin permissions.
Anyway if user run Umit from tarball it should be possible because the necessary configs are ready to access.

How to create empty directories with distutils?
I was reading this thread:
http://mail.python.org/pipermail/distutils-sig/2009-January/010773.html

It is not possible yet. This patch overwride make_distribution() of distutils.sdist and then it trigger in the middle a function to create it. Once Python support it make_distribution() overwride should be removed.


git-svn-id: http://svn.umitproject.org/svnroot/umit/trunk@5323 1105ee14-b0fa-0310-85a5-ff3eed429ff7